# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotSetupRequest {
  /** Telegram bot token from BotFather */
  token: string;
  /** Numeric Telegram user ID of the admin */
  adminId: number;
  /** Max file size in MB (default 50) */
  maxFileSizeMb?: number;
  /** Max requests per minute per user (default 5) */
  rateLimitRequests?: number;
}

export interface BotSetupResponse {
  success: boolean;
  botUsername?: string;
  botId?: number;
  message: string;
}

export interface BotStatus {
  running: boolean;
  setupDone: boolean;
  botUsername?: string;
  adminId?: number;
  pid?: number;
}

export interface BotActionResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}