# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/setup:
    post:
      operationId: setupBot
      tags: [bot]
      summary: Configure and start the Telegram bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotSetupRequest"
      responses:
        "200":
          description: Bot set up and started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotSetupResponse"
        "400":
          description: Invalid credentials or setup error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    BotSetupRequest:
      type: object
      properties:
        token:
          type: string
          description: Telegram bot token from BotFather
        adminId:
          type: number
          description: Numeric Telegram user ID of the admin
        maxFileSizeMb:
          type: number
          description: Max file size in MB (default 50)
        rateLimitRequests:
          type: number
          description: Max requests per minute per user (default 5)
      required:
        - token
        - adminId
    BotSetupResponse:
      type: object
      properties:
        success:
          type: boolean
        botUsername:
          type: string
        botId:
          type: number
        message:
          type: string
      required:
        - success
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error