# lib/api-zod/src/generated/types/botSetupRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotSetupRequest {
  /** Telegram bot token from BotFather */
  token: string;
  /** Numeric Telegram user ID of the admin */
  adminId: number;
  /** Max file size in MB (default 50) */
  maxFileSizeMb?: number;
  /** Max requests per minute per user (default 5) */
  rateLimitRequests?: number;
}