# lib/api-zod/src/generated/types/botSetupResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotSetupResponse {
  success: boolean;
  botUsername?: string;
  botId?: number;
  message: string;
}